Asset pipeline first pass

+++ src/graphic/common/visualnode.h
@@ -26,8 +26,8 @@ struct VisualNodeData
     const void* m_VertexBuffer;
     const void* m_IndexBuffer;
 
-    uint16_t m_NumVertices;
-    uint16_t m_NumIndices;
+    uint32_t m_NumVertices;
+    uint32_t m_NumIndices;
 
     const ethXMMatrix* m_ModelMatrix;
 };
@@ -40,12 +40,12 @@ public:
 
     inline D3D12_VERTEX_BUFFER_VIEW GetVertexBufferView() const { return m_VertexBufferView; }
     inline D3D12_INDEX_BUFFER_VIEW GetIndexBufferView() const { return m_IndexBufferView; }
-    inline uint16_t GetNumIndices() const { return m_StaticData.m_NumIndices; }
+    inline uint32_t GetNumIndices() const { return m_StaticData.m_NumIndices; }
     inline const ethXMMatrix GetModelMatrix() const { return *m_StaticData.m_ModelMatrix; }
 
 private:
-    void UploadVertexBuffer(const void* data, uint16_t numVertices);
-    void UploadIndexBuffer(const void* data, uint16_t numIndices);
+    void UploadVertexBuffer(const void* data, uint32_t numVertices);
+    void UploadIndexBuffer(const void* data, uint32_t numIndices);
 
     void InitVertexBufferView(size_t bufferSize, size_t stride);
     void InitIndexBufferView(size_t bufferSize);
